refactor: breaking dependecy graph cycle
- move `periodBlockRange` from helpers to Period (so Period doesn't depend on helpers)
- move `calcInputs` and `calcOutputs` from helpers to Transaction (so Transaction doesn't depend on helpers)
