+++ lib/helpers.js
@@ -7,22 +7,9 @@
  */
 
 import { bufferToHex } from 'ethereumjs-util';
-import {
-  BigInt,
-  add,
-  subtract,
-  greaterThan,
-  greaterThanOrEqual,
-  lessThan,
-  equal,
-} from 'jsbi-utils';
 import Outpoint from './outpoint';
-import Input from './input';
-import Output from './output';
 import Period from './period';
 import Tx from './transaction';
-
-import { BLOCKS_PER_PERIOD } from './constants';
 
 function getLowerCase(value) {
   if (value) {
@@ -170,76 +157,25 @@ export function extendWeb3(web3Instance) {
   return web3Instance;
 }
 
+/**
+ * DEPRECATED: Use Transaction.calcInputs instead
+ */
 export function calcInputs(unspent, from, amount, color) {
-  const myUnspent = unspent.filter(
-    ({ output }) =>
-      output.color === color &&
-      output.address.toLowerCase() === from.toLowerCase()
-  );
-
-  const exact = myUnspent.find(utxo =>
-    equal(BigInt(utxo.output.value), BigInt(amount))
-  );
-
-  if (exact) return [new Input(exact.outpoint)];
-
-  const inputs = [];
-  let sum = BigInt(0);
-  for (let i = 0; i < myUnspent.length; i += 1) {
-    inputs.push(new Input(myUnspent[i].outpoint));
-    sum = add(sum, BigInt(myUnspent[i].output.value));
-
-    if (greaterThanOrEqual(sum, BigInt(amount))) {
-      break;
-    }
-  }
-
-  if (lessThan(sum, BigInt(amount))) {
-    throw new Error('Not enough inputs');
-  }
-
-  return inputs;
-}
-
-// ToDo: handle inputs from different accounts
-// ToDo: handle different input colors
-export function calcOutputs(unspent, inputs, from, to, amount, color) {
-  if (unspent.length === 0) {
-    throw new Error('Unspent is empty');
-  }
-
-  const inInputs = u =>
-    inputs.findIndex(input => u.outpoint.equals(input.prevout)) > -1;
-  const sum = unspent
-    .filter(inInputs)
-    .reduce((a, u) => add(a, BigInt(u.output.value)), BigInt(0));
-
-  if (lessThan(sum, BigInt(amount))) {
-    throw new Error('Not enough inputs');
-  }
-
-  const outputs = [new Output(amount, to.toLowerCase(), color)];
-  if (greaterThan(sum, BigInt(amount))) {
-    outputs.push(
-      new Output(subtract(sum, BigInt(amount)), from.toLowerCase(), color)
-    );
-  }
-
-  return outputs;
+  return Tx.calcInputs(unspent, from, amount, color);
 }
 
 /**
- * Returns the block number interval for the period given block is included to.
- *
- * @param {Number} blockNumber block height of the block we are getting period block range for
- * @returns {Array} block interval in [startBlock: number, endBlock: number] format
+ * DEPRECATED: Use Transaction.calcOutputs instead
+ */
+export function calcOutputs(unspent, inputs, from, to, amount, color) {
+  return Tx.calcOutputs(unspent, inputs, from, to, amount, color);
+}
+
+/**
+ * DEPRECATED: Use Period.periodBlockRange instead
  */
 export function periodBlockRange(blockNumber) {
-  const periodNum = Math.floor(blockNumber / BLOCKS_PER_PERIOD);
-  return [
-    periodNum * BLOCKS_PER_PERIOD,
-    (periodNum + 1) * BLOCKS_PER_PERIOD - 1,
-  ];
+  return Period.periodBlockRange(blockNumber);
 }
 
 /**
