+++ lib/period.js
@@ -11,7 +11,8 @@ import { bufferToHex, toBuffer, keccak256 } from 'ethereumjs-util';
 import MerkleTree from './merkleTree';
 import Block from './block';
 import Util from './util';
-import { periodBlockRange } from './helpers';
+
+import { BLOCKS_PER_PERIOD } from './constants';
 
 export default class Period {
   constructor(prevHash, blocks) {
@@ -132,6 +133,19 @@ export default class Period {
     return proof;
   }
 
+  /**
+   * Returns the block number interval for the period given block is included to.
+   *
+   * @param {Number} blockNumber block height of the block we are getting period block range for
+   * @returns {Array} block interval in [startBlock: number, endBlock: number] format
+   */
+  static periodBlockRange(blockNumber) {
+    const periodNum = Math.floor(blockNumber / BLOCKS_PER_PERIOD);
+    return [
+      periodNum * BLOCKS_PER_PERIOD,
+      (periodNum + 1) * BLOCKS_PER_PERIOD - 1,
+    ];
+  }
 
  /**
  * Creates new Period with all the blocks from the given interval.
@@ -163,7 +177,7 @@ export default class Period {
  */
   static periodForTx(plasma, tx) {
     const { blockNumber } = tx;
-    const [startBlock, endBlock] = periodBlockRange(blockNumber);
+    const [startBlock, endBlock] = Period.periodBlockRange(blockNumber);
     return Period.periodForBlockRange(plasma, startBlock, endBlock);
   }
