+++ lib/transaction.js
@@ -7,9 +7,15 @@
  */
 
 import ethUtil from 'ethereumjs-util';
-import { BigInt } from 'jsbi-utils';
-
-import { calcInputs, calcOutputs } from './helpers';
+import {
+  BigInt,
+  add,
+  subtract,
+  greaterThan,
+  greaterThanOrEqual,
+  lessThan,
+  equal,
+} from 'jsbi-utils';
 import Input, { SPEND_INPUT_LENGTH } from './input';
 import Output, { OUT_LENGTH, NST_OUT_LENGTH } from './output';
 import Outpoint from './outpoint';
@@ -90,14 +96,14 @@ export default class Transaction {
   }
 
   static transferFromUtxos(utxos, from, to, value, color) {
-    const inputs = calcInputs(
+    const inputs = Transaction.calcInputs(
       utxos,
       from,
       value,
       color
     );
 
-    const outputs = calcOutputs(
+    const outputs = Transaction.calcOutputs(
       utxos,
       inputs,
       from,
@@ -621,6 +627,64 @@ export default class Transaction {
         throw new Error(`unknown transaction type: ${type}.`);
       }
     }
+  }
+
+  static calcInputs(unspent, from, amount, color) {
+    const myUnspent = unspent.filter(
+      ({ output }) =>
+        output.color === color &&
+        output.address.toLowerCase() === from.toLowerCase()
+    );
+  
+    const exact = myUnspent.find(utxo =>
+      equal(BigInt(utxo.output.value), BigInt(amount))
+    );
+  
+    if (exact) return [new Input(exact.outpoint)];
+  
+    const inputs = [];
+    let sum = BigInt(0);
+    for (let i = 0; i < myUnspent.length; i += 1) {
+      inputs.push(new Input(myUnspent[i].outpoint));
+      sum = add(sum, BigInt(myUnspent[i].output.value));
+  
+      if (greaterThanOrEqual(sum, BigInt(amount))) {
+        break;
+      }
+    }
+  
+    if (lessThan(sum, BigInt(amount))) {
+      throw new Error('Not enough inputs');
+    }
+  
+    return inputs;
+  }
+  
+  // ToDo: handle inputs from different accounts
+  // ToDo: handle different input colors
+  static calcOutputs(unspent, inputs, from, to, amount, color) {
+    if (unspent.length === 0) {
+      throw new Error('Unspent is empty');
+    }
+  
+    const inInputs = u =>
+      inputs.findIndex(input => u.outpoint.equals(input.prevout)) > -1;
+    const sum = unspent
+      .filter(inInputs)
+      .reduce((a, u) => add(a, BigInt(u.output.value)), BigInt(0));
+  
+    if (lessThan(sum, BigInt(amount))) {
+      throw new Error('Not enough inputs');
+    }
+  
+    const outputs = [new Output(amount, to.toLowerCase(), color)];
+    if (greaterThan(sum, BigInt(amount))) {
+      outputs.push(
+        new Output(subtract(sum, BigInt(amount)), from.toLowerCase(), color)
+      );
+    }
+  
+    return outputs;
   }
 
 }
